fix: rename file hideSegment.ts to maskSegment.ts
