+++ src/transformations/maskSegment.ts
